Update conkyrc
There are sadly two things in this conkyrc update. First is the port
number for outbound has been increased by one. This is to remove some
false positives that I was seeing. Second is the moving of the
prime-select intel vs nvidia usage line. I thought it was more useful
somewhere else, so I moved it.